新增前端/ros端坐标转换方法、延长了后端sleep时间、修改Point相关位置的属性类型为Float

+++ backend/src/main/java/com/se/hw/controller/RosController.java
@@ -172,7 +172,7 @@ public class RosController {
 
     @GetMapping("/getLocation")
     public Result getLocation() {
-        HashMap<String, Double> axis = new HashMap<>();
+        HashMap<String, Float> axis = new HashMap<>();
         axis.put("xAxis", RosGlobal.point.getXAxis());
         axis.put("yAxis", RosGlobal.point.getXAxis());
         return Result.success(100, axis);
